+++ rubti_beancount_import/tests/spk_giro/test_data.csv.beancount
